Fix vbrsender: make Python checks actually get compressed
NewPythonCheckLoader calls collectoraggregator.InitializeCheckContext once,
at loader-construction time, storing the manager in a package-level global;
SubmitMetric always resolves a check's sender through that captured global,
ignoring whichever manager pkg/collector/scheduler.go's getChecks() picked
per-config. Selecting between two different manager instances upstream (one
plain, one VBR-wrapped) therefore silently never took effect for Python —
the real openmetrics check would have shipped uncompressed with zero
telemetry if ever allowlisted.

Move the compress-or-not decision inside vbrsender.SenderManager.GetSender
itself: it now checks the check name against the vbr_compression_checks
allowlist per call, wrapping and caching only when it matches, and passing
straight through to the inner manager otherwise (which already caches per
ID on its own). This makes vbrsender.Wrap's manager the only manager
instance in the system — Python's captured global now captures the
VBR-aware one too, with no Python-side code changes needed.

pkg/collector/scheduler.go collapses to a single senderManager field and
drops the per-config branch in getChecks() accordingly.

Found via automated review on PR #53123.
